move more classes to packages
