+++ UploadDaemon/Upload/FileSystemUpload.cs
@@ -1,12 +1,7 @@
 using System;
-using System.Collections.Generic;
 using System.IO;
 using System.IO.Abstractions;
-using System.Net;
-using System.Net.Http;
-using System.Text;
 using System.Threading.Tasks;
-using System.Web;
 using NLog;
 using UploadDaemon.SymbolAnalysis;
