+++ UploadDaemon/UploadDaemon.csproj
@@ -106,8 +106,8 @@
     <Compile Include="Archiving\IDateTimeProvider.cs" />
     <Compile Include="Configuration\Config.cs" />
     <Compile Include="Configuration\ConfigParser.cs" />
-    <Compile Include="LineCoverageMerger.cs" />
-    <Compile Include="MessageFormatter.cs" />
+    <Compile Include="Upload\LineCoverageMerger.cs" />
+    <Compile Include="Utils\MessageFormatter.cs" />
     <Compile Include="SymbolAnalysis\FileCoverage.cs" />
     <Compile Include="Archiving\PurgeArchiveTask.cs" />
     <Compile Include="SymbolAnalysis\ILineCoverageSynthesizer.cs" />
@@ -115,9 +115,9 @@
     <Compile Include="SymbolAnalysis\LineCoverageSynthesizer.cs" />
     <Compile Include="SymbolAnalysis\ParsedTraceFile.cs" />
     <Compile Include="SymbolAnalysis\SymbolCollection.cs" />
-    <Compile Include="UploadTask.cs" />
-    <Compile Include="TraceFile.cs" />
-    <Compile Include="TraceFileScanner.cs" />
+    <Compile Include="Upload\UploadTask.cs" />
+    <Compile Include="Scanning\TraceFile.cs" />
+    <Compile Include="Scanning\TraceFileScanner.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Configuration\AzureFileStorage.cs" />
     <Compile Include="Upload\HttpClientUtils.cs" />
